example/requirements.txt
attrs==21.4.0
black==22.3.0
click==8.1.2
mypy==0.942
mypy-extensions==0.4.3
numpy==1.22.3
pathspec==0.9.0
platformdirs==2.5.1
scipy==1.8.0
tomli==2.0.1
typing-extensions==4.1.1